Update JavaTimer class to guard against premature timer expiration
The JavaTimer class uses Thead.sleep() to wait a specified number of
milliseconds. Thread.sleep() requires the caller catch
InterruptedException, but in the original implementation, such an
exception would result in the code assuming that the specified time
had been reached.

So as things stood, if you e.g. set a timer for 60 seconds, then the
timer might generate its "onTick" signal before that 60 seconds had
elapsed.

This patch changes the code so that the method will wait until the
specified time has been reached.

The "milliseconds" parameters are also changed to "long", which is the
type used by the rest of the java library for millisecond values.

Added a bit of javadoc and a toString() method as well.

Note there is still a "FIXME" in the code which I've not addressed.

Test-information:

Tested in debugging setup; things seem to be working as expected.

+++ src/com/isode/stroke/network/JavaTimerFactory.java
@@ -4,7 +4,7 @@
  * See Documentation/Licenses/GPLv3.txt for more information.
  */
 /*
- * Copyright (c) 2010, Isode Limited, London, England.
+ * Copyright (c) 2010-2012, Isode Limited, London, England.
  * All rights reserved.
  */
 
@@ -18,7 +18,7 @@ public class JavaTimerFactory implements TimerFactory {
         eventLoop_ = eventLoop;
     }
 
-    public Timer createTimer(int milliseconds) {
+    public Timer createTimer(long milliseconds) {
         return new JavaTimer(eventLoop_, milliseconds);
     }
